+++ src/Makefile.am
@@ -109,6 +109,7 @@ BITCOIN_CORE_H = \
   netbase.h \
   net.h \
   noui.h \
+  policy/estimator.h \
   policy/feerate.h \
   policy/policy.h \
   primitives/block.h \
@@ -137,6 +138,7 @@ BITCOIN_CORE_H = \
   tinyformat.h \
   txdb.h \
   txmempool.h \
+  txmempoolentry.h \
   ui_interface.h \
   uint256.h \
   undo.h \
@@ -186,6 +188,7 @@ libbitcoin_server_a_SOURCES = \
   miner.cpp \
   net.cpp \
   noui.cpp \
+  policy/estimator.cpp \
   policy/feerate.cpp \
   policy/policy.cpp \
   rest.cpp \
@@ -199,6 +202,7 @@ libbitcoin_server_a_SOURCES = \
   timedata.cpp \
   txdb.cpp \
   txmempool.cpp \
+  txmempoolentry.cpp \
   validationinterface.cpp \
   $(JSON_H) \
   $(BITCOIN_CORE_H)
